+++ js/get-random-id.js
@@ -1,0 +1,19 @@
+import {getRandomInteger} from './get-random-element.js';
+
+function createRandomIdFromRangeGenerator (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+
+    previousValues.push(currentValue);
+
+    return currentValue;
+  };
+}
+
+export {createRandomIdFromRangeGenerator};
